simplified a lot code
Lost genericity in favour of clarity, and more than that, it seems to
work now. We use views for creating table body as it is much much
simpler. Note that facility column is taking  common_name, this will be
fixed. Genericity loss is explained by the fact that the webdesigner
must choose which columns will be displayed, the site won't decide it
on its own.